fix(Toast): fixes after review
