+++ packages/orbit-components/src/Toast/__tests__/index.test.jsx
@@ -1,14 +1,16 @@
 // @flow
 import * as React from "react";
-import { render, screen, fireEvent } from "@testing-library/react";
+import { render, screen, fireEvent, waitFor } from "@testing-library/react";
+import userEvent from "@testing-library/user-event";
 
 import { Airplane } from "../../icons";
 import Toast from "../Toast";
 
 describe("Toast", () => {
-  it("should have expected DOM output", () => {
+  it("should have expected DOM output", async () => {
     const onMouseEnter = jest.fn();
     const onMouseLeave = jest.fn();
+    const onDismiss = jest.fn();
 
     render(
       <Toast
@@ -17,7 +19,8 @@ describe("Toast", () => {
         onMouseLeave={onMouseLeave}
         ariaLive="polite"
         role="status"
-        onDismiss={() => {}}
+        visible
+        onDismiss={onDismiss}
         placement="bottom-center"
       >
         kek
@@ -26,11 +29,17 @@ describe("Toast", () => {
 
     const toast = screen.getByText("kek");
 
-    fireEvent.mouseOver(toast);
+    userEvent.hover(toast);
     expect(onMouseEnter).toHaveBeenCalled();
 
-    fireEvent.mouseLeave(toast);
+    userEvent.unhover(toast);
     expect(onMouseLeave).toHaveBeenCalled();
+
+    fireEvent.mouseDown(toast, { screenX: 10 });
+    fireEvent.mouseMove(toast, { screenX: 300 });
+    fireEvent.mouseUp(toast);
+
+    await waitFor(() => expect(onDismiss).toHaveBeenCalled(), { timeout: 1000 });
 
     expect(screen.getByTestId("airplane")).toBeInTheDocument();
     expect(screen.getByRole("status")).toHaveAttribute("aria-live", "polite");
