+++ packages/orbit-components/src/Toast/index.jsx
@@ -1,7 +1,7 @@
 // @flow
 import { toast } from "react-hot-toast";
 
-import ToastInit from "./ToastInit";
+import ToastProvider from "./ToastProvider";
 
 import type { createToastType, createToastPromiseType } from ".";
 
@@ -9,5 +9,5 @@ const createToast: createToastType = (content, { icon }) => toast(content, { ico
 const createToastPromise: createToastPromiseType = (content, { icon, loading, error, success }) =>
   toast.promise(content, { icon, loading, error, success });
 
-export default ToastInit;
+export default ToastProvider;
 export { createToast, createToastPromise };
