+++ packages/orbit-components/src/utils/rtl/index.js
@@ -43,6 +43,7 @@ export const translate3d: Translate3d = value => ({ theme }) => {
   if (!theme.rtl) {
     return `translate3d(${value})`;
   }
+
   const parts = value.split(",").filter(part => !Number.isNaN(parseFloat(part)) && part);
   const x = parts[0];
   const newX = x[0] === "-" ? x.slice(1) : `-${x}`;
